Make the clear button be styled correctly for the enhanced select element
enhanced select elements have a wrapper div added to the dom by accessible-autocomplete, this causes issues with our current css implementation for the clear button, which did not work with an extra wrapper div.
I've updated the css to expect an extra wrapper div and updated the enhanced text input mode to add the same wrapper div that the enhaced select mode has, this way they have the same dom structure after accessible-autocomplete is executed.