Remove the `o-autocomplete__container` element.
To reduce code complexity and css specificity.

The child `o-autocomplete__wrapper` element can be used instead.
Since it is created by the alphagov dependency there's an increased
risk of conflicts but we are already styling elements created by the
dependency, we can mitigate this by pinning the dependency version
number.