Provide an API for applications to be able to transform the return values from their `source` function before rendering the suggestions.
The use-case that this is required for, is when the `source` function is returning data that is not an array of strings, such as an array of objects. For example the objects could contain an ID field and a string field so that they can more easily track what suggestion was selected.